backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
bcrypt==4.0.1
PyJWT==2.8.0
cryptography==41.0.4
python-dotenv==1.0.0
marshmallow==3.20.1
Pillow==10.0.1
Werkzeug==2.3.7
stripe==6.6.0
email-validator==2.0.0
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
gunicorn==21.2.0
